Add deleteSongFromPlaylist service incomplete

+++ src/components/Song.jsx
@@ -2,10 +2,12 @@ import React, { useState, useEffect } from "react";
 import { RiMusic2Line, RiEditBoxLine, RiDeleteBin2Line } from "react-icons/ri";
 import { deleteSong, editSong } from "../services/SongService";
 import EditSongForm from "./EditSongForm";
+import { deleteSongFromPlaylist } from "../services/SongXPlaylist";
 
-function Song({ title, duration }) {
+function Song({ title, duration, code }) {
   const [showIcons, setShowIcons] = useState(false);
   const [showEditForm, setShowEditForm] = useState(false);
+  const [showDeleteForm, setShowDeleteForm] = useState(false);
 
   function onMouseOver() {
     setShowIcons(true);
@@ -14,8 +16,12 @@ function Song({ title, duration }) {
     setShowIcons(false);
   }
 
-  function Delete() {
-    deleteSong(title);
+  function Delete(typeDelete) {
+    if (typeDelete === 1) {
+      deleteSong(title);
+    } else if (typeDelete === 2) {
+      // deleteSongFromPlaylist() //TODO: obtener codigo de la cancion
+    }
   }
 
   function showForm() {
@@ -25,6 +31,10 @@ function Song({ title, duration }) {
   const handleCloseForm = () => {
     setShowEditForm(false);
   };
+
+  function showFormDelete() {
+    setShowDeleteForm(true);
+  }
 
   const handleSubmitForm = (data) => {
     editSong(title, data.newTitle, data.newDuration);
@@ -36,7 +46,7 @@ function Song({ title, duration }) {
       onMouseOver={onMouseOver}
       onMouseOut={onMouseLeave}
     >
-      <div class="flex">
+      <div className="flex">
         <RiMusic2Line className="text-blue-500 text-2xl mr-2" />
         <div>
           <h2 className="text-lg font-bold">{title}</h2>
@@ -44,19 +54,44 @@ function Song({ title, duration }) {
         </div>
       </div>
       {showIcons && (
-        <div class="flex">
+        <div className="flex">
           <RiEditBoxLine
             className="text-blue-500 text-2xl mr-2"
             onClick={showForm}
           />
           <RiDeleteBin2Line
             className="text-blue-500 text-2xl mr-2"
-            onClick={Delete}
+            onClick={showFormDelete}
           />
         </div>
       )}
       {showEditForm && (
         <EditSongForm onClose={handleCloseForm} onSubmit={handleSubmitForm} />
+      )}
+      {showDeleteForm && (
+        <div className="fixed z-50 top-0 left-0 w-screen h-screen flex items-center justify-center bg-bg-light bg-opacity-40">
+          <h1 className="text-purple-dark text-xl font-bold">Elimnar cancion</h1>
+          <div className="bg-white p-4 rounded shadow flex flex-col">
+            <button
+              className=" px-8 py-2 rounded"
+              onClick={() => Delete(1)} // Envolver console.log en una función de flecha
+            >
+              Borrar permanentemente
+            </button>
+            <button
+              className="px-8 py-2 rounded"
+              onClick={() => Delete(2)} // Envolver console.log en una función de flecha
+            >
+              Borrar de la playlist
+            </button>
+            <button
+              className="bg-bg-light text-white px-8 py-2 rounded"
+              onClick={() => setShowDeleteForm(false)}
+            >
+              Cancelar
+            </button>
+          </div>
+        </div>
       )}
     </div>
   );
